spring-boot-security-course/src/main/java/com/example/demo/employee/EmployeeController.java
package com.example.demo.employee;

import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api")
@CrossOrigin(origins = "localhost:3000")
public class EmployeeController {
    @Autowired
    EmployeeRepository employeeRepository;

    @GetMapping("/get")
    public List<EmployeeEntity> getAll() {
        return employeeRepository.findAll();
    }

    // TODO: Implement a POST request route that accepts data to save to the employees table

}
